Create README.md
Deep Tech like metaverse is my interest area. The use of blockchain technology is going to change the existing concepts. The decentralization of assets is the option that is coming late but for sure. Cryptos and NFTs will be the virtual transacting media that will rule real-world transactions. The concepts will confuse the convention economy so much that there will be a huge hue and cry in many corners of the globe. But, get ready. I will do my bit to alert the super minds.